Typing fix.

+++ robotpy_toolkit_7407/sensors/photonvision/photon_camera.py
@@ -3,7 +3,7 @@ from robotpy_toolkit_7407.sensors.photonvision.photon_target import PhotonTarget
 from wpimath.geometry import Pose3d
 
 class PhotonCamera:
-    def __init__(self, name: str, poseRelativeToRobot: Pose3d, scale_constant: int = 1, height=None, pitch=None):
+    def __init__(self, name: str, poseRelativeToRobot: Pose3d, scale_constant: float = 1, height=None, pitch=None):
         self.camera = photonvision.PhotonCamera(cameraName=name)
         self.latest_best_target: PhotonTarget = None
         self.latest_targets_all: list[PhotonTarget] = None
